restructured all files
