+++ src/components/main/PrivacyAndPlicy.jsx
