+++ src/components/main/Reviews.jsx
@@ -2,14 +2,14 @@ import React from 'react'
 import { FaQuoteLeft, FaStar, FaStarHalfAlt } from 'react-icons/fa'
 import { Swiper, SwiperSlide } from 'swiper/react'
 import { Pagination, Autoplay, Navigation } from 'swiper/modules'
-import studentPic from '../assets/images/student1.jpg'
-import studentPic1 from '../assets/images/student2.jpg'
-import studentPic2 from '../assets/images/student3.jpg'
-import studentPic3 from '../assets/images/student4.jpg'
-import studentPic4 from '../assets/images/student7.jpg'
-import studentPic5 from '../assets/images/student6.jpg'
-import studentPic6 from '../assets/images/student9.jpg'
-import studentPic7 from '../assets/images/student10.jpg'
+import studentPic from '../../assets/images/student1.jpg'
+import studentPic1 from '../../assets/images/student2.jpg'
+import studentPic2 from '../../assets/images/student3.jpg'
+import studentPic3 from '../../assets/images/student4.jpg'
+import studentPic4 from '../../assets/images/student7.jpg'
+import studentPic5 from '../../assets/images/student6.jpg'
+import studentPic6 from '../../assets/images/student9.jpg'
+import studentPic7 from '../../assets/images/student10.jpg'
 
 function Reviews() {
  return (
