+++ src/components/main/index.js
@@ -1,0 +1,11 @@
+export { default as Contact } from "../../pages/contact/Contact";
+export { default as Home } from "../../pages/home/Home";
+export { default as About } from "../../pages/about/About";
+export { default as Footer } from "./Footer";
+export { default as PrivacyAndPlicy } from "./PrivacyAndPlicy";
+export { default as TermsAndCondPopup } from "./TermsAndCondPopup";
+export { default as Testimonial } from "./Testimonial";
+export { default as ForgetPassword } from "../auth/ForgetPassword";
+export { default as Login } from "../auth/Login";
+export { default as SignUp } from "../auth/SignUp";
+export { default as Navbar } from "../navbar/Navbar";
